+++ .github/workflows/go.yml
@@ -7,10 +7,15 @@ jobs:
     steps:
       - uses: actions/checkout@master
 
-      - name: Run test
+      - name: Run unit tests
         uses: cedrickring/golang-action@1.4.1
         with:
           args: make unit-test
+
+      - name: Run Go lint
+        uses: cedrickring/golang-action@1.4.1
+        with:
+          args: make lint
 
       - name: Bulid binary
         uses: cedrickring/golang-action@1.4.1
