+++ pkg/inputs/inputs_flag_test.go
@@ -1,17 +1,20 @@
-// Copyright © 2019 Thilina Manamgoda
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
+/*
+ * Copyright (c) 2019 WSO2 Inc. (http:www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http:www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
 
 package inputs
 
@@ -20,25 +23,6 @@ import (
 	"gotest.tools/assert"
 	"testing"
 )
-
-func TestIsPasswordValid(t *testing.T) {
-	t.Run("SuccessTest", testIsPasswordValidSuccessFunc())
-	t.Run("FailedTestEmptyPassword", testIsPasswordValidFailedEmptyFunc())
-}
-
-func testIsPasswordValidSuccessFunc() func(t *testing.T) {
-	return func(t *testing.T) {
-		result := IsPasswordValid("password")
-		assert.Equal(t, true, result)
-	}
-}
-
-func testIsPasswordValidFailedEmptyFunc() func(t *testing.T) {
-	return func(t *testing.T) {
-		result := IsPasswordValid("")
-		assert.Equal(t, false, result)
-	}
-}
 
 func TestGetFlagBoolVal(t *testing.T) {
 	t.Run("SuccessTest", testGetFlagBoolValSuccessFunc())
@@ -122,44 +106,21 @@ func testGetFlagStringArrayValFailedFunc() func(t *testing.T) {
 	}
 }
 
-func TestIsValidSingleArg(t *testing.T) {
-	t.Run("SuccessTest", testIsValidSingleArgSuccessFunc())
-	t.Run("FailedTestWithEmptyArgs", testIsValidSingleArgEmptyArgFailedFunc())
-	t.Run("FailedTestWithMultipleArgs", testIsValidSingleArgMultipleArgsFailedFunc())
+func TestIsPasswordValid(t *testing.T) {
+	t.Run("SuccessTest", testIsPasswordValidSuccessFunc())
+	t.Run("FailedTestEmptyPassword", testIsPasswordValidFailedEmptyFunc())
 }
 
-func testIsValidSingleArgSuccessFunc() func(t *testing.T) {
+func testIsPasswordValidSuccessFunc() func(t *testing.T) {
 	return func(t *testing.T) {
-		assert.Equal(t, true, IsValidSingleArg([]string{"a"}))
+		result := IsPasswordValid("password")
+		assert.Equal(t, true, result)
 	}
 }
 
-func testIsValidSingleArgEmptyArgFailedFunc() func(t *testing.T) {
+func testIsPasswordValidFailedEmptyFunc() func(t *testing.T) {
 	return func(t *testing.T) {
-		assert.Equal(t, false, IsValidSingleArg([]string{}))
+		result := IsPasswordValid("")
+		assert.Equal(t, false, result)
 	}
 }
-
-func testIsValidSingleArgMultipleArgsFailedFunc() func(t *testing.T) {
-	return func(t *testing.T) {
-		assert.Equal(t, false, IsValidSingleArg([]string{"a", "b"}))
-	}
-}
-
-func TestIsArgValid(t *testing.T) {
-	t.Run("SuccessTest", testIsArgValidSuccessFunc())
-	t.Run("FailedTest", testIsArgValidFailedFunc())
-}
-
-func testIsArgValidSuccessFunc() func(t *testing.T) {
-	return func(t *testing.T) {
-		assert.Equal(t, true, IsArgValid("test"))
-	}
-}
-
-func testIsArgValidFailedFunc() func(t *testing.T) {
-	return func(t *testing.T) {
-		assert.Equal(t, false, IsArgValid(""))
-	}
-}
-
